Update Thompson Portal Automation.py
changed the way I find the pdf for Lee County. I now just grab the link that contains today's date. As long as it has today's date in the href, whatever else they call it doesn't matter. Because of this, I moved the error handling for the pdf out of the move function and into the section where we try to download the file.